requirements.txt
streamlit>=1.28.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.8.0
email-validator>=2.1.0.post1
python-dotenv>=1.0.0


